No response View Added Notif For Geofencing QR Scanning and Generating

+++ app/src/main/java/com/altwav/samquicksal2/notificationsActivity/DeclinedNotificationActivity.kt
@@ -44,6 +44,6 @@ class DeclinedNotificationActivity : AppCompatActivity() {
             }
         })
 
-        viewModel.getNotifDeclinedInfo(cust_id, notif_id.toInt())
+        viewModel.getNotifDeclinedInfo(cust_id, notif_id)
     }
 }
